+++ src/main/java/com/ssh/lotto/LottoResultReviewer.java
@@ -4,11 +4,11 @@ import java.util.Map;
 
 public class LottoResultReviewer {
 
-    static public double getReturnRate(int money, Map<LottoWinningEnum, Integer> lottoWinningMap) {
+    static public int getReturnRate(int money, Map<LottoWinningEnum, Integer> lottoWinningMap) {
         int sum = 0;
         for (LottoWinningEnum lottoAwardEnum : lottoWinningMap.keySet()) {
             sum += lottoAwardEnum.getPrizeMoney() * lottoWinningMap.get(lottoAwardEnum);
         }
-        return sum * 1.0 / money * 1.0 * 100;
+        return (int) (sum * 1.0 / money * 1.0 * 100);
     }
 }
